Added dynamic waypoint navigation
If waypoint array in Patrol.cs is empty, a list of objects tagged "Waypoint" will be queried. The closest "Waypoint" obj that is not the current or 2nd-to-last target will be selected as a new target. If the current and 2nd-to-last "Waypoint" objs are the only 2, it will pick the 2nd-to-last. No target will be set if the current target is the only "Waypoint" obj.